Rename all files and test class for consistency
