+++ src/usethis/_interface/tool.py
@@ -52,11 +52,30 @@ def codespell(
 
 
 @app.command(
-    name="coverage",
+    name="coverage.py",
     help="Use Coverage.py: a code coverage measurement tool.",
     rich_help_panel="Testing",
 )
 def coverage_py(
+    remove: bool = remove_opt,
+    how: bool = how_opt,
+    offline: bool = offline_opt,
+    quiet: bool = quiet_opt,
+    frozen: bool = frozen_opt,
+) -> None:
+    with (
+        usethis_config.set(offline=offline, quiet=quiet, frozen=frozen),
+        files_manager(),
+    ):
+        _run_tool(use_coverage_py, remove=remove, how=how)
+
+
+@app.command(
+    name="coverage",
+    help="Use Coverage.py: a code coverage measurement tool.",
+    rich_help_panel="Testing",
+)
+def coverage(
     remove: bool = remove_opt,
     how: bool = how_opt,
     offline: bool = offline_opt,
