+++ src/usethis/_tool/all_.py
@@ -1,6 +1,6 @@
 from usethis._tool.base import Tool
 from usethis._tool.impl.codespell import CodespellTool
-from usethis._tool.impl.coverage import CoveragePyTool
+from usethis._tool.impl.coverage_py import CoveragePyTool
 from usethis._tool.impl.deptry import DeptryTool
 from usethis._tool.impl.import_linter import ImportLinterTool
 from usethis._tool.impl.pre_commit import PreCommitTool
