+++ tests/usethis/_tool/impl/test_coverage_py.py
@@ -3,10 +3,10 @@ from pathlib import Path
 from usethis._config_file import files_manager
 from usethis._integrations.file.pyproject_toml.io_ import PyprojectTOMLManager
 from usethis._test import change_cwd
-from usethis._tool.impl.coverage import CoveragePyTool
+from usethis._tool.impl.coverage_py import CoveragePyTool
 
 
-class TestCoverageTool:
+class TestCoveragePyTool:
     class TestAddConfigs:
         def test_after_codespell(self, tmp_path: Path):
             # To check the config is valid
